requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
pyjwt==2.10.1
email-validator==2.3.0
pytest==8.4.1
httpx==0.28.1
